+++ index.js
@@ -86,6 +86,28 @@ module.exports = {
     ],
 
     // *** Different from Airbnb
+    // Excludes for..of - I'm using redux-saga in my project and it's the only readable way
+    // to put multiple actions, had similar experience in async/await
+    'no-restricted-syntax': [
+      'error',
+      {
+        selector: 'ForInStatement',
+        message:
+          'for..in loops iterate over the entire prototype chain, which is virtually never what you want. Use Object.{keys,values,entries}, and iterate over the resulting array.'
+      },
+      {
+        selector: 'LabeledStatement',
+        message:
+          'Labels are a form of GOTO; using them makes code confusing and hard to maintain and understand.'
+      },
+      {
+        selector: 'WithStatement',
+        message:
+          '`with` is disallowed in strict mode because it makes code impossible to predict and optimize.'
+      }
+    ],
+
+    // *** Different from Airbnb
     // Disabled because it's tedious when using with Redux's mapDispatchToProps
     // https://eslint.org/docs/rules/no-shadow
     'no-shadow': 'off',